+++ src/main/java/podo/odeego/web/auth/jwt/dto/GenerateTokenResponse.java
@@ -1,4 +1,4 @@
-package podo.odeego.web.security.jwt.dto;
+package podo.odeego.web.auth.jwt.dto;
 
 public record GenerateTokenResponse (
 	String accessToken,
